+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 2.0.0 (2024-10-03)
+## Features
+- Added .NET 8.0 support.
+- Added `Redis.Cache.Testing` package for testing extensions.
+
+## Breaking Changes
+- Renamed package to `Redis.Cache` from `Redis.Sidecar.Cache`.
+- Dropped support for .NET 6.0.
+- Swapped extensions for `WebApplicationBuilder` for extensions for `ServiceCollection`.
+
 # 1.3.0 (2024-01-12)
 ## Features
 - Added ability to specify configuration settings during instantiation of `RedisCache`